# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "convert pdf to img(jpg)",
        "title": "pdf2img",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
        "/upload/pdf": {
            "post": {
                "description": "get pdf",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "upload"
                ],
                "summary": "recieve pdf from FrontEnd",
                "parameters": [
                    {
                        "type": "file",
                        "description": "\t",
                        "name": "pdf",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handler.ResponseInfo": {
            "type": "object",
            "properties": {
                "baseName": {
                    "type": "string"
                },
                "count": {
                    "type": "integer"
                }
            }
        },
        "handler.SuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "info": {
                    "$ref": "#/definitions/handler.ResponseInfo"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    }
}